side running, no database output

+++ src/main/java/at/itkolleg/fullstackserver1/repository/IAuthorRepository.java
@@ -3,6 +3,6 @@ package at.itkolleg.fullstackserver1.repository;
 import at.itkolleg.fullstackserver1.model.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IAuthorRepository extends JpaRepository<Author, Integer> {
+public interface IAuthorRepository extends JpaRepository<Author, Long> {
 
 }
