+++ src/main/java/at/itkolleg/fullstackserver1/repository/ICitationRepository.java
@@ -3,5 +3,6 @@ package at.itkolleg.fullstackserver1.repository;
 import at.itkolleg.fullstackserver1.model.Citation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICitationRepository extends JpaRepository<Citation, Integer> {
+public interface ICitationRepository extends JpaRepository<Citation, Long> {
+
 }
